// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/Enumerations.ts
export enum SoeEntityState { 
    Active = 0,
    Inactive = 1,
    Deleted = 2,
    Temporary = 3,
    Hidden = 4
}
// ...
export enum TermGroup_DataStorageRecordAttestStatus { 
    None = 0,
    Initialized = 1,
    PartlySigned = 2,
    Signed = 3,
    Rejected = 4,
    Cancelled = 5
}
// ...
export enum XEMailAnswerType { 
    None = 0,
    Yes = 1,
    No = 2
}