// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/SOECompModelDTOs.ts
    export interface IDataStorageRecordDTO {
        dataStorageRecordId: number;
        recordId: number;
        entity: SoeEntityType;
        attestStateId?: number;
        currentAttestUsers: string;
        attestStatus: TermGroup_DataStorageRecordAttestStatus;
        data: number[];
        type: SoeDataStorageRecordType;
        attestStateName: string;
        attestStateColor: string;
        roleIds: number[];
    }

    

    

    export interface IDataStorageRecipientDTO {
        dataStorageRecipientId: number;
        dataStorageId: number;
        userId: number;
        readDate?: Date;
        confirmedDate?: Date;
        state: SoeEntityState;
        userName: string;
        employeeNrAndName: string;
    }
// ...
    export interface IDocumentDTO {
        dataStorageId: number;
        messageId?: number;
        userId?: number;
        name: string;
        description: string;
        fileName: string;
        extension: string;
        fileSize?: number;
        folder: string;
        validFrom?: Date;
        validTo?: Date;
        readDate?: Date;
        confirmedDate?: Date;
        needsConfirmation: boolean;
        answerType: XEMailAnswerType;
        answerDate?: Date;
        created?: Date;
        createdBy: string;
        modified?: Date;
        modifiedBy: string;
        records: IDataStorageRecordDTO[];
        recipients: IDataStorageRecipientDTO[];
        messageGroupIds: number[];
        attestStatus: TermGroup_DataStorageRecordAttestStatus;
        currentAttestUsers: string;
        attestStateId?: number;
        displayName: string;
    }